demo/requirements.txt
streamlit>=1.30.0
numpy>=1.24.0
plotly>=5.18.0
Pillow>=10.0.0
torch>=2.0.0
umap-learn>=0.5.0
safetensors>=0.4.0
